requirements.txt
py-clob-client>=0.34.5
python-dotenv==1.2.1
pandas==2.3.3
sortedcontainers==2.4.0
eth-account==0.13.7
eth-utils==5.3.1
poly_eip712_structs==0.0.1
py_order_utils==0.3.2
requests==2.32.5
websockets==15.0.1
cryptography==46.0.3
web3==7.14.0
